requirements.txt
# LangChain core
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-huggingface>=0.0.3

# Vector store
faiss-cpu>=1.7.4

# Google AI
google-generativeai>=0.8.0

# Document parsing
pdfplumber>=0.11.0
pandas>=2.1.0
openpyxl>=3.1.0
json-repair>=0.30.0

# Image processing (required for scanned PDF OCR)
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Dashboard
dash>=2.14.0
plotly>=5.18.0

# API
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Embeddings
sentence-transformers>=2.2.0